mod_md: fixed a conversion warning, bumped version number.
git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/trunk@1888726 13f79535-47bb-0310-9956-ffa450edef68

+++ modules/md/mod_md_ocsp.c
@@ -159,7 +159,7 @@ int md_ocsp_get_stapling_status(unsigned char **pder, int *pderlen,
     rv = md_ocsp_get_status(ocsp_copy_der, &ctx, sc->mc->ocsp, &id, c->pool, md);
     if (APR_STATUS_IS_ENOENT(rv)) goto declined;
     *pder = ctx.der;
-    *pderlen = ctx.der_len;
+    *pderlen = (int)ctx.der_len;
     return OK;
     
 declined:
